Implemented avl tree class with working insertItem(), rotateLeft() and rotateRight() methods. the insert item inherits the same code from the BSTree class, however at the end it checks the balance factor, which is just the difference between the left and right subtrees of any given node, and then if the balance factor of a node is 2 or greater then the tree will be rotated left, and if less then -2, rotate right.
rotateLeft() works by first creating temporary nodes for the root,, tree.right and tree.right.left.
in order to rotate, tree.right needs to become the new root
the old root needs to move to the left of the new root.
the old tree.right.left is smaller than tree.right, and therefore smaller than the new root, but larger than the old root, so it needs to go to the right of the old root